fixed enclave test compilation
