+++ cosmwasm/enclaves/query/src/lib.rs
@@ -4,8 +4,6 @@ extern crate sgx_tstd as std;
 
 extern crate sgx_types;
 
-mod tests;
-
 use std::env;
 
 #[allow(unused_imports)]
@@ -66,84 +64,4 @@ fn set_log_level_or_default(default: LevelFilter, max_level: LevelFilter) {
     }
 
     log::set_max_level(log_level);
-}
-
-#[cfg(feature = "test")]
-pub mod logging_tests {
-    use crate::{count_failures, set_log_level_or_default};
-    use ctor::*;
-    use lazy_static::lazy_static;
-    use log::*;
-    use log::{Metadata, Record};
-    use std::sync::SgxMutex;
-    use std::{env, panic};
-
-    lazy_static! {
-        static ref LOG_BUF: SgxMutex<Vec<String>> = SgxMutex::new(Vec::new());
-    }
-    pub struct TestLogger;
-    impl log::Log for TestLogger {
-        fn enabled(&self, _metadata: &Metadata) -> bool {
-            true
-        }
-        fn log(&self, record: &Record) {
-            LOG_BUF.lock().unwrap().push(format!(
-                "{}  [{}] {}",
-                record.level(),
-                record.target(),
-                record.args()
-            ));
-        }
-        fn flush(&self) {}
-    }
-
-    #[ctor]
-    fn init_logger_test() {
-        log::set_logger(&TestLogger).unwrap();
-    }
-
-    pub fn run_tests() {
-        println!();
-        let mut failures = 0;
-
-        count_failures!(failures, {
-            test_log_level();
-            test_log_default_greater_than_max();
-        });
-
-        if failures != 0 {
-            panic!("{}: {} tests failed", file!(), failures);
-        }
-    }
-
-    fn test_log_level() {
-        env::set_var("LOG_LEVEL", "WARN");
-        set_log_level_or_default(LevelFilter::Error, LevelFilter::Info);
-        assert_eq!(log::max_level(), LevelFilter::Warn);
-        info!("Should not process");
-        assert!(LOG_BUF.lock().unwrap().is_empty());
-
-        env::set_var("LOG_LEVEL", "TRACE");
-        set_log_level_or_default(LevelFilter::Error, LevelFilter::Info);
-        assert_eq!(log::max_level(), LevelFilter::Error);
-        debug!("Should not process");
-        assert!(LOG_BUF.lock().unwrap().is_empty());
-
-        env::set_var("LOG_LEVEL", "WARN");
-        set_log_level_or_default(LevelFilter::Warn, LevelFilter::Warn);
-        assert_eq!(log::max_level(), LevelFilter::Warn);
-        trace!("Should not process");
-        assert!(LOG_BUF.lock().unwrap().is_empty());
-
-        warn!("This should process");
-        assert_eq!(LOG_BUF.lock().unwrap().len(), 1);
-    }
-
-    fn test_log_default_greater_than_max() {
-        eprintln!("The following should fail:");
-        let result = panic::catch_unwind(|| {
-            set_log_level_or_default(LevelFilter::Trace, LevelFilter::Error);
-        });
-        assert!(result.is_err());
-    }
 }
