+++ cosmwasm/enclaves/shared/contract-engine/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [features]
 query-only = []
 debug-print = []
+test = []
 
 # This annotation is here to trick the IDE into showing us type information about this crate.
 # We always compile to the "sgx" target, so this will always be false.
