+++ cosmwasm/enclaves/shared/cosmos-types/Cargo.toml
@@ -4,6 +4,9 @@ version = "1.2.4"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
 
+[features]
+test = []
+
 # This annotation is here to trick the IDE into showing us type information about this crate.
 # We always compile to the "sgx" target, so this will always be false.
 # when compiling to the "sgx" target, we pull this from the target root with an "extern crate" directive
