+++ cosmwasm/enclaves/shared/cosmos-types/src/types.rs
@@ -506,59 +506,3 @@ impl SignerInfo {
         Ok(signer_info)
     }
 }
-
-#[cfg(feature = "test")]
-pub mod tests {
-
-    use super::*;
-    // use crate::crypto::{AESKey, SIVEncryptable, Seed, KEY_MANAGER};
-
-    // todo: fix test vectors to actually work
-    pub fn test_new_from_slice() {
-        let nonce = [0u8; 32];
-        let user_public_key = [0u8; 32];
-        let msg = "{\"ok\": \"{\"balance\": \"108\"}\"}";
-
-        let mut slice = nonce.to_vec();
-        slice.extend_from_slice(&user_public_key);
-        slice.extend_from_slice(msg.as_bytes());
-
-        let secret_msg = SecretMessage {
-            nonce,
-            user_public_key,
-            msg: msg.as_bytes().to_vec(),
-        };
-
-        let msg_from_slice = SecretMessage::from_slice(&slice).unwrap();
-
-        assert_eq!(secret_msg, msg_from_slice);
-    }
-
-    // This is commented out because it's trying to modify KEY_MANAGER which is immutable.
-    // // todo: fix test vectors to actually work
-    // pub fn test_msg_decrypt() {
-    //     let seed = Seed::new().unwrap();
-    //
-    //     KEY_MANAGER
-    //         .set_consensus_seed(seed)
-    //         .expect("Failed to set seed");
-    //
-    //     let nonce = [0u8; 32];
-    //     let user_public_key = [0u8; 32];
-    //
-    //     let msg = "{\"ok\": \"{\"balance\": \"108\"}\"}";
-    //     let key = calc_encryption_key(&nonce, &user_public_key);
-    //
-    //     let encrypted_msg = key.encrypt_siv(msg.as_bytes(), &[&[]]);
-    //
-    //     let secret_msg = SecretMessage {
-    //         nonce,
-    //         user_public_key,
-    //         msg: encrypted_msg,
-    //     };
-    //
-    //     let decrypted_msg = secret_msg.decrypt()?;
-    //
-    //     assert_eq!(decrypted_msg, msg)
-    // }
-}
