+++ cosmwasm/enclaves/shared/crypto/src/lib.rs
@@ -31,30 +31,3 @@ pub use ed25519::{Ed25519PublicKey, KeyPair, PUBLIC_KEY_SIZE, SECRET_KEY_SIZE};
 
 pub use hash::sha::{sha_256, HASH_SIZE};
 pub use traits::{Encryptable, Hmac, Kdf, SIVEncryptable, SealedKey, HMAC_SIGNATURE_SIZE};
-
-#[cfg(feature = "test")]
-pub mod tests {
-    use super::*;
-    use crate::count_failures;
-
-    pub fn run_tests() {
-        println!();
-        let mut failures = 0;
-
-        count_failures!(failures, {
-            // kdf::tests::test_derive_key();
-            // storage::tests::test_open();
-            // storage::tests::test_seal();
-            multisig::tests_decode_multisig_signature::test_decode_sig_sanity();
-            multisig::tests_decode_multisig_signature::test_decode_long_leb128();
-            multisig::tests_decode_multisig_signature::test_decode_wrong_long_leb128();
-            multisig::tests_decode_multisig_signature::test_decode_malformed_sig_only_prefix();
-            multisig::tests_decode_multisig_signature::test_decode_sig_length_zero();
-            multisig::tests_decode_multisig_signature::test_decode_malformed_sig_wrong_length();
-        });
-
-        if failures != 0 {
-            panic!("{}: {} tests failed", file!(), failures);
-        }
-    }
-}
